refactor(st-09058): modularize tool lifecycle
