+++ packages/core/src/tools/lifecycle-error.ts
@@ -1,0 +1,3 @@
+export function getErrorMessage(error: unknown): string {
+  return error instanceof Error ? error.message : String(error);
+}
