+++ planning/features/09-solid-micro-refactors-feature-plan.md
@@ -2,8 +2,8 @@
 
 **Epic Range:** EP-09 through EP-09
 **Status:** In Progress
-**Last Updated:** 2026-06-05
-**Active Story:** None currently (post-merge sync complete)
+**Last Updated:** 2026-06-06
+**Active Story:** ST-09058 - Modularize Core Tool Lifecycle and Tests (In Review)
 
 ---
 
@@ -99,7 +99,7 @@ Recent improvement snapshot:
 - `ST-09055` merged on 2026-06-04 after reducing `packages/tools/src/data/relational/schema/schema-inspector.ts` from `725` lines to a `126` line public facade, extracting focused shared, PostgreSQL, MySQL, and SQLite runtime helpers while replacing the schema-inspector test monolith with focused PostgreSQL, cache, and filter suites and keeping the explicit-`any` baseline flat at `workspace 84/289` and `tools 53/67`.
 - `ST-09056` merged on 2026-06-05 after reducing `packages/skills/src/registry.ts` from `506` lines to a `101` line public facade, extracting focused discovery, event, prompt, query, and shared-type helpers while replacing the `419` line registry test monolith with focused discovery, query, event, and rescan suites. The story preserved `SkillRegistry` public behavior, absorbed review follow-ups for temp-dir cleanup, shared-type alignment, and documentation accuracy, and kept the explicit-`any` baseline flat at `workspace 84/289` and `skills 0/0`.
 - `ST-09061` through `ST-09066` were added on 2026-06-05 to keep EP-09 stocked beyond the current lane with another modularization-first batch, prioritizing large runtime files in `core` and `patterns` before the remaining schema-hardening slices so future fixes land on smaller modules instead of monoliths.
-- `EP-09` remains open as the daily hardening stream, with `ST-09057` now merged, the current ready lane centered on `ST-09058` and `ST-09061` through `ST-09063`, and `ST-09064` through `ST-09066` plus the schema-tightening `ST-09059` and `ST-09060` queued behind them in backlog order.
+- `EP-09` remains open as the daily hardening stream, with `ST-09058` now in review, the remaining ready lane centered on `ST-09061` through `ST-09063`, and `ST-09064` through `ST-09066` plus the schema-tightening `ST-09059` and `ST-09060` queued behind them in backlog order.
 - The refreshed follow-on queue now extends beyond the current Ready lane so another few weeks of small SOLID/DRY and modularization work can be pulled without re-planning the epic.
 
 ---
