+++ planning/kanban-queue.md
@@ -1,12 +1,12 @@
 # Kanban Queue: AgentForge
 
-**Last Updated:** 2026-06-05
+**Last Updated:** 2026-06-06
 
 ## Queue Status Summary
 
-- **Ready:** 4 stories
+- **Ready:** 3 stories
 - **In Progress:** 0 stories
-- **In Review:** 0 stories
+- **In Review:** 1 story
 - **Blocked:** 0 stories
 - **Backlog:** 5 stories
 
@@ -14,7 +14,6 @@
 
 ## Ready
 
-- `ST-09058` - Modularize Core Tool Lifecycle and Tests
 - `ST-09061` - Modularize Core Tool Types and Tests
 - `ST-09062` - Modularize Core Tool Executor and Tests
 - `ST-09063` - Modularize Multi-Agent Worker Node and Tests
@@ -28,7 +27,7 @@ None currently.
 
 ## In Review
 
-None currently.
+- `ST-09058` - Modularize Core Tool Lifecycle and Tests
 
 ## Blocked
